renamed package namespace into edu.northeastern, signed APK
