+++ app/build.gradle
@@ -3,17 +3,25 @@ plugins {
 }
 
 android {
-    namespace 'com.ebookfrenzy.numad23sp_team16'
+    signingConfigs {
+        team16ReleaseConfig {
+            storePassword 'GoodluckNEU2023!'
+            keyAlias 'groupKey'
+            keyPassword 'GoodluckNEU2023!'
+        }
+    }
+    namespace 'edu.northeastern.numad23sp_team16'
     compileSdk 33
 
     defaultConfig {
-        applicationId "com.ebookfrenzy.numad23sp_team16"
+        applicationId "edu.northeastern.numad23sp_team16"
         minSdk 26
         targetSdk 33
         versionCode 1
         versionName "1.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        signingConfig signingConfigs.team16ReleaseConfig
     }
 
     buildTypes {
