--- app/src/main/java/com/ebookfrenzy/numad23sp_team16/MainActivity.java
@@ -1,20 +1,0 @@
-package com.ebookfrenzy.numad23sp_team16;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-    }
-
-    public void startWebServiceActivity(View view) {
-        startActivity(new Intent(MainActivity.this, WebServiceActivity.class));
-    }
-}
