--- app/src/main/java/com/ebookfrenzy/numad23sp_team16/Name.java
@@ -1,13 +1,0 @@
-package com.ebookfrenzy.numad23sp_team16;
-
-public class Name {
-    String name;
-
-    public Name(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
