+++ app/src/main/java/edu/northeastern/numad23sp_team16/Name.java
@@ -1,0 +1,13 @@
+package edu.northeastern.numad23sp_team16;
+
+public class Name {
+    String name;
+
+    public Name(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
